Add Turtle class and implement associated behavior
A new Turtle class is added to the Animals within the Organism hierarchy. This class implements a custom 'step' method with 75% probability of ignoring the move. This new class is integrated within the existing application code, referenced in 'main.cpp' and included in the 'CMakeLists.txt'. Changes in 'World.cpp' accommodate the Turtle's unique movement behavior.